agent-api/requirements.txt
requests
flask
kafka-python
protobuf
flask_socketio
pandas